tiles.center.explore looks as intended

+++ src/tile2net/tiles/minibatch.py
@@ -388,7 +388,7 @@ class MiniBatch(
     @look_at(tile2net.tileseg.inference.inference.Inference.validate)
     def submit_polygons(self):
         affines = next(self.tiles.stitched.affine_iterator())
-        arrays = to_numpy(self.predictions)
+        arrays = to_numpy(self.predictions).astype(np.uint8)
         files = next(self.tiles.outdir.polygons.iterator())
         it = zip(arrays, affines, files)
         for array, affine, file in it:
